+++ NWN.Anvil.Tests/src/main/API/Object/NwPlaceableTests.cs
@@ -176,7 +176,7 @@ namespace Anvil.Tests.API
     }
 
     [TearDown]
-    public void CleanupTestObject()
+    public void CleanupTestObjects()
     {
       foreach (NwGameObject testObject in createdTestObjects)
       {